Make sure blockdevs are actually sorted on return from get_blockdevs
Recently it began to seem desirable that they be sorted in the same
order as the metadata. At that time, I inspected the code and concluded
mistakenly that they were sorted and updated the specification to indicate
this. Now, they are actually sorted, which is not necessary for
correctness in current setup, but may be later.

Signed-off-by: mulhern <amulhern@redhat.com>