Abstract the action of setting up a pool
Make pool setting up behavior uniform in StratEngine::initialize()
and StratEngine::block_evaluate(). Previously, initialize() would return
an error on a duplicate pool name, now it persists, just like
block_evaluate.

Check the metadata for a name conflict before settng up a StratPool.
This way, it is not necessary to turn around and tear the pool right
back down again if there is a name conflict.

Consider a name conflict to be warning-, not error-worthy. In the unlikely
event that this happens, it could be the result of stale metadata in the
partial pool being processed.

* Make a new method, setup_pool, which either sets up a pool or returns an
error.
* Use this method in StratEngine::initialize() and
StratEngine::block_evaluate()
* This makes some methods dead in the source but still live in the tests.
Annotate them appropriately.
* Tighten up is_stratis_device() header comment. The plan is to change
its spec so that it also returns a StaticHeader, so that reading the
metadata on the devices is not repeated unnecessarily.
* Make a separate method in pool module to check metadata for consistency.
* Change StratPool::setup() so that it takes a metadata parameter and
change calling sites appropriately.
* Update tests.

Signed-off-by: mulhern <amulhern@redhat.com>