Get rid of a bunch of check() functions that aren't used anymore
These function were made dead by PR#703 and previous PRs which introduced
an eventing mechanism supplanting the code that called check() on the
engine every few seconds. They are no longer needed. This change also
eliminates two FIXMEs and the need for some other code improvements having
to do with the instantiation of a devicemapper context.

Signed-off-by: mulhern <amulhern@redhat.com>