Change Eventable and get_evented() signature
get_evented() returns a reference to a static Eventable.
This way it can make use of the get_dm_context() method rather than
calling DM::new().

Eventable methods do not take mut self parameters. That this is possible
is due to the fact that there is only one implementation of Eventable,
the one for the strat engine, and that this eventable does not need to
be mutated for any actions to be performed on it, because the mutation
all happens in the outside world. So, in a general pursuit of least
strength, it makes sense to drop the mut modifiers.

Signed-off-by: mulhern <amulhern@redhat.com>