Make self parameters to Eventable methods non-mut
They do not need to take a mutable argument, because the DM context
is never mutated and a DM context is the one thing that implements
Eventable in the stratisd code base.

As a consequence it is possible to make Engine::eventable()
return an immutable object, so do that.

Remove some mut modifiers where the devicemapper context is used as
an Eventable.

Signed-off-by: mulhern <amulhern@redhat.com>