Pass devices by value to get_blockdevs
This way, the key and values stored in devices can just be moved out of
devices into the StratBlockDev that is created at the end.
The cost of all this is that the error message in pool_setup() must be
constructed regardless of whether it is used, or, must be constructed
at every point where it is found that a message needs to be constructed.

It really does seem more desirable to pass devices as a reference and clone
the individual values when extracted.

Signed-off-by: mulhern <amulhern@redhat.com>