Omit pool UUID field from ThinPool and BlockDevMgr
If the goal is to store only the mapping of a UUID to a pool and not to
store the UUID in the pool, having the UUID in structs which are
owned by the pool is a bad idea. It doesn't readily lead to errors,
because once set the UUID can not change, but it is inconsistant with the
whole approach.

Signed-off-by: mulhern <amulhern@redhat.com>