Adjust cleanup code in various ways
Have one_iteration() chain an error if list_devices() fails.
It better explains the problem than previous summary error, which indicated
a failure to remove devices.

Wrap actions in an anonymous function to make one summary exit error.

Use chain_err() instead of with_chain() in some places.
This is possible when the error itself is an error-chain Error.

Iterate over the first element of the tuple returnd by list_devices.
The tuple doesn't need to be carried around everywhere if the name is
all that gets used.

Signed-off-by: mulhern <amulhern@redhat.com>