metadata: Use sync_all() instead of flush()
Apparently flush() does nothing, so use sync_all() instead to ensure
data is saved.

Since we also want our functions to be generic so we can mock using Cursor,
Add trait SyncAll, which allows Cursor to still work with the new method
call (sync_all() on a Cursor is no-op.)

There were also a couple spots that were taking a reference that didn't
need to, that needed to be corrected to get this working.

fixes #892

Signed-off-by: Andy Grover <agrover@redhat.com>