Try to copy back old data if initialize fails
Starting in initialize() and proceeding into callees that write data,
read the existing data first and store it, along with where it goes, into
a RestoreState struct. In the event an error occurs before successful
completion of initialize(), attempt to write back the pre-existing data.

Much of this work is in order to share F, which is either a File or a
memory Cursor (used for testing.) Wrap it in a Rc<RefCell>>. This is mostly
driven by Cursor, which unlike File does not have a cloneable handle.

fixes #937

Signed-off-by: Andy Grover <agrover@redhat.com>