Attach setting blockdev id to Pool trait
Previously it was attached to the BlockDev trait.
This eliminates the mistake of calling save_state in the D-Bus layer.

Remove save_state from Pool trait because it is now unnecesary.
Remove get_mut_blockdev from Pool trait because it is now unnecessary.
Both these methods should continue unnecessary, as any other methods that
might change blockdevs, should be attached to the Pool in order to handle
the possible necessity of having to update pool metadata.

Signed-off-by: mulhern <amulhern@redhat.com>