Have method in StratEngine implementation return StratBlockDevs
Previously, these methods returned BlockDevs. At the internal implementation
level, it makes more sense to return objects of the concrete type, rather
than trait objects.

This means that, instead of being cast to BlockDev at the lowest level, these
references are cast to BlockDev at the highest level, i.e., in the impl
of the Pool trait for StratPool.

Signed-off-by: mulhern <amulhern@redhat.com>