Rewrite get_devices() and slice_devices()
* Remove size_ok and slice_count. These two methods find the size of a
devices and then perform one other action. Instead of doing this, the
code now reads the device size just once and then hangs on to it for the
necessary calculations.
* Add more precise documentation to slice_disk and slice_devices().
* Change slice_devices() so that it takes a Vec of pairs, of which the
second element is the size of the device.
* Remove a comparision in slice_devices() that doesn't work in the general
case: finding the min of the number of chunks available on a single disk and
the total number of chunks required. This comparison only makes sense when
there is only one device to be chunked.
* Teardown any excess linear devs in order to get exactly the amount requested
in slice_devices().
* In get_devices(), just make min_size always 1 GiB if not set. Document
the restriction that occurs when it is necessary to partition devices w/
a FIXME.

Signed-off-by: mulhern <amulhern@redhat.com>