Have strat engine verify existance of required binaries
Assume that all binaries will be required by stratisd at some time,
and that it shouldn't run if they are not available.

This seems better to me than having a panic if we try to use them and they
are not needed.

It looks like using lazy_static and using the techniques that were used
for implementing get_dm are not really all that different.

Hoist cmd.rs up a level, since it seems like the engine really has to
interact with it.

Signed-off-by: mulhern <amulhern@redhat.com>