Avoid unwrap() in get_dm() method
It seemed that it was necessary to hav an unwrap() in get_dm() because it
seemed that it was necessary to have get_dm() return a simple type, i.e.,
one without an Option or Result. The problem with this approach is that
it required the use of catch_unwind() at the point that get_dm() was
first called, to ensure an orderly exit.

This solution is exactly equivalent, but leaves out the unwind() and
hence the requirement to use catch_unwrap().

It introduces a new function, get_dm_init() which returns a StratisResult.
It changes get_dm() so that it invokes get_dm_init() but immediately
expects a non-error value.

The solution is semantically equivalent to the previous approach; under
all circumstances it should behave in an equivalent manner.

Signed-off-by: mulhern <amulhern@redhat.com>