Strip message argument out of execute_cmd
It's not required for actually executing the method.

Instead, include the command that failed in the error message.
The Debug representation of the command that failed will show the
executable and the arguments exactly.

Also, construct an error if attempting to execute the command fails.
Leave out the check in get_executable, as it is redundant.

We should plan that, in future the two distinct error strings returned by
execute_cmd will be two programmatically distinct errors, that ancillary code
can inspect further to discover the most likely cause of the errors
it sees.

Signed-off-by: mulhern <amulhern@redhat.com>