Use ftruncate instead of fallocate to make loopbacked devices
Because ftruncate makes a sparse file, this helps to avoid running out of
space on the device which backs the loopbacked devices.

The downside is that some failures may yet occur but be postponed until
the devices are actually being filled, i.e., while the test is actually
being run.

Don't wipe the first 1 MiB, with ftruncate, the device should present as
zeroed.

Signed-off-by: mulhern <amulhern@redhat.com>