Add a comment explaining why polling is done with libc
Otherwise, somebody might come along and try to change it so it's using
nix and come to an unexpected impasse near the end of the process.
Why would someone want to change polling to use nix? So, that poll would
return a Result, and there wouldn't have to be special treatment of the
error code from libc::poll.

Signed-off-by: mulhern <amulhern@redhat.com>