Ensure BDA copies are intact
When we read up BDA from a device we will verify both copies and correct
in those cases where we have one good copy and one invalid copy.  This
change also corrects a bug where we would return no Stratis BDA if the
magic on the first copy contains an error even though we have a complete
and valid second copy.

Signed-off-by: Tony Asleson <tasleson@redhat.com>