Use udev properties to identify Stratis devices
* Change the notion of how ownership of devices is determined.
Previously ownership was determined exclusively by in-Stratis judgements.
Now, some of the identification task is allocated to udev.

Previously there were three ownership possibilities, Ours, Theirs, and Unowned.
Now there are four, Ours, Theirs, Unowned, and Contradiction.
Contradiction can arise if udev believes that the device belongs to Stratis
but Stratis can not identify it as such. Consistently return an error
when a contradiction is found.

Enrich the Theirs constructor with some information about how the device
has been identified as not belonging to Stratis.

Eliminate determine_ownership() method entirely. This method's way of
identifying a device belonging to another system was to check the first
few sectors to see if it was dirty. This was very rough, and it makes more
sense to rely on libblkid, which should be able to id block devices, as that
is its job.

Update tests that use determine_ownership() to use either device_identifiers
or identify(), two new methods added by this PR.

Omit a test that expected that checking for dirty bits was the correct
way to determine if a device belonged to another system.

* Add new method device_identifiers(). This is a simple auxiliary method
that just reads the Stratis static header from a device and returns Stratis
pool UUID and device UUID, None if Stratis header was not found, and an
error if there was an error making the determination. This method is a
module level method of the metadata module, as it gets its information
solely from Stratis metadata.

* Add new method identify(). This is a method that uses udev information
as well as Stratis metadata to identify a single device. It is located
in the device module in backstore, because it is not solely associated
with metadata.

* Get rid of method is_stratis_device() entirely. This method originated
from a previous approach to identifying devices which relied on filesystem
data and Stratis metadata entirely.

* Rewrite find_all to work just like identify(). Here, the order of
operations is different, since udev is used to scan all devices, but the
effect should be identical to discovering all devices via a udev add event
and then identifying them using the identify method.

* For both identify and find_all have a clearly set off chunk of code
to handle the fall back if an older version of libblkid that can not
identify Stratis devices is running.

* To support the operation of both find_all and identify add an
unclaimed() method which checks some sort of udev magic to determine
whether the device is considered by udev to be unclaimed by any other
system.