Revert "Leverage the udev db"
This reverts commit d6948e16527f0a30b84ac2624653833e7de4a2e9.

This also removes a test that would be removed in a subsequent commit in
PR and hacks up device_ownership (which will also be removed in this PR)
in order to make it compile.

This also adds back in cmd::udev_settle(), since it is already in use in
related code.

It makes BDA_STATIC_HDR_SIZE temporarily public in order to make
test_force_flag_dirty compile.

At this stage, it is not expected that all tests will pass, although they
might.