Initial buff_log implementation
buff_log wraps another logger and buffers it instead of outputting it,
until the dump() command is used. This allows logging to be verbose, but
only outputted if an error is encountered.

Documented, but I was having a hard time coming up with tests, since
logged messages are not accessible for making comparisons.

Signed-off-by: Andy Grover <agrover@redhat.com>