Use buff-log in stratisd
If not executed with '--debug', use buff-log to hold back log output unless
explicitly dumped. This can occur if a dbus method call fails, or if some
other error causes the event loop to terminate.

Also, set a timer to periodically wake up to record the current state of
the engine to the log. This is pretty verbose but will only be output if
something goes wrong so it's ok!

Getting this working required supporting a signalfd, and the use of
alarm(2), which is not currently supported by Nix (but we should work on
getting it added.) This is because if somebody hits ctrl-C we want to exit
gracefully so we have a chance to print the buffered log.

Signed-off-by: Andy Grover <agrover@redhat.com>