Separate buffered logger specific configuration from rest
Some of the configuration is done at the request of the user, but some
is done in order to modify the log entries of the buffered logger so that
they are correct when finally output. Call these operations, really just
one at this time, out explicitly and explain their purpose.

Signed-off-by: mulhern <amulhern@redhat.com>