Move setup_pool() out of setup
It's better to consider it an auxiliary engine method than a method
associated with the backstore because it doesn't just get devices,
it calls StratPool::setup.

This way we export one fewer methods from backstore, which is good, because
encapsulation is increased, and we have fewer reverse dependencies in
setup.rs on StratPool.

Signed-off-by: mulhern <amulhern@redhat.com>