Tidy up the body of the loop
Do not spend an identifier on actual_size.
Its value is calculated only to verify that size hasn't decreased. It is not
used in the rest of the method, so it doesn't need a name.

Add more detail to several error messages.

Comment some things a bit more thoroughly.

Make metadata locating block more rustic. This makes it seem a bit clearer
to me.

Signed-off-by: mulhern <amulhern@redhat.com>