Have create_dbus_pool take a Pool
This way, the newly created object path can be set in the create_dbus_pool
method. I believe that this will always be necessary, so it makes sense to
combine the two actions.

This also eliminates the possibility of mistakenly passing the wrong object
path to the set_dbus_path method.

Signed-off-by: mulhern <amulhern@redhat.com>