Remove hard coded indices in event loop
Instead of utilizing fixed positions for specific file descriptor
types, lets utilize another vector which contains the type of fd
at that index.  This way we can place the fd in any index and still
process them.  This will allow us to add new fd types easier and
reduce the burden of keeping track of what goes where.

Additionally this change removes some fds churn because everytime we
went through the loop for any reason we would remove and add the dbus
pollfd structures.  This is only needed when we actually process
dbus file descriptors.

Signed-off-by: Tony Asleson <tasleson@redhat.com>