Remove restriction about either cache or block device existing
This is still true, but assertions and comments that are false are
removed and some methods are altered.

The biggest change is that Backstore::device() must return an
Option<Device>, the expect that was previously in Backstore::device()
must be moved out to every invocation site and a context-specific message
inserted there.