Add PoolState
An embryonic effort to track the state of the pool. There are a class of
errors that are severe, but not so severe that stratisd should terminate.
Having a Bad pool state may be a nice way to handle these, where a severe
error causes the pool to enter this state, and then maybe it can recover
back to Good (or not), but we can then perhaps use the state to avoid
further harmful actions, as well a notifying the user that a problem has
occurred.

Signed-off-by: Andy Grover <agrover@redhat.com>