Make notify_listeners not depend on if there's a dbus object path
I think the intention was that the listener model should be general
purpose(?), and not just for bubbling up DBus events. It's fine to
include the dbus path in the notification, but whether the notification is
issued or not should not depend on the path existing.

Since, yes, in this case we *do* need the path, the check for a dbus path
moves to the listener code in stratisd.rs.

Also, make EngineEvent include refs, it doesn't appear to need to be owned.

Signed-off-by: Andy Grover <agrover@redhat.com>