Change get_mount_point() into mount_points()
Returns a Vec, which may be empty.

This is a more capable function. The previous behavior can be
emulated by using the .first() method.

One other change that was made was to be less tolerant of errors: any parse
error will cause the method to fail. This should never happen, since
we're parsing a file whose fields are defined and fixed as part of a kernel
API.

Signed-off-by: Andy Grover <agrover@redhat.com>