Borrow, instead of clone, connection wherever possible
It's a general goal to limit the number of copies of a pointer we have
floating around to the number necessary. This brings them down to 2, just
one in the run() method in stratisd, and one in the EventHandler struct.

This way, the borrows are not buried in invoked methods, but all up-front
in the run() method.

It turns out that the Connection impl never requires a mutable self
argument. That means it is never necessary to do a mutable borrow of a
Connection. But that should not cause any feeling of security, because most
of those methods do something interesting to their internal D-Bus
connection, which is mutable. This is what one would expect, as most of
these methods clearly can not operate w/out changing something.

Signed-off-by: mulhern <amulhern@redhat.com>