Move generic event stuff into a generic module
Export it without cfg flag, because it is really generic, and not specific
to D-Bus.

The reason to move event stuff out of engine proper is that everything in
engine is a trait that must be implemented elsewhere, where as the
engine stuff is a generic implementation, like the Table, that can
be shared by both implementations.

Signed-off-by: mulhern <amulhern@redhat.com>