Remove support for filesystem.MountPoints from API and fs trait
We've decided not to show it in the cli, because it's already accessible
via the `mount` command. Given that, it doesn't appear to be useful or
used, so why support the method? We can always add it later, after all.

StratEngine support for getting mount points is still needed, so it
moves from the trait impl to the regular impl block.

Signed-off-by: Andy Grover <agrover@redhat.com>