Get rid of some comments
Someday we will change DATA_BLOCK_SIZE or learn to do something else wierd
but the comment will stick around forever.

Metablocks is more stable, but developers need to know that it is stable,
so best to leave it out as well.

Also, the addition of the comments really didn't belong in the commit
where it ended up.

Signed-off-by: mulhern <amulhern@redhat.com>