Further parameterize call to extend_thin_sub_device()
Instead of taking &mut self, change extend_thin_sub_device to take
&mut ThinPoolDev and &mut Vec of existing segs. This lets us pass
different values for existing_segs from the data/meta wrapping functions,
which means that the common function can operate without differentiating
between the two use cases.

Another change that this requires is not using self.backstore_device, but
instead using backstore.device(), which is ok since they must be the same.

Finally, this calls thinpooldev.suspend() instead of thinpool.suspend().
The only difference is the latter also suspends the MDV, but this
should be ok since the MDV is not in the thinpool.

Signed-off-by: Andy Grover <agrover@redhat.com>