Check free space and calc lowater even if not extending data device
These should be called every time check() is called, not just when we
are extending the data dev. For example, a meta extension could cause the
pool to change PoolSpaceState, or if the data dev is fully extended we
still want those to be run when the event is triggered.

Signed-off-by: Andy Grover <agrover@redhat.com>