Add PoolState and PoolSpaceState
Add PoolState and FreeSpaceState members to struct ThinPool. PoolState is
mostly a placeholder for future work, such as #378 or #1122.

FreeSpaceState is used to mark transitions based upon percentage used vs.
total available where the pool is getting close to full. It uses
throttling in the Warn state to ensure the daemon has a chance to respond
and enter the Crit state before the pool actually runs out of free
DataBlocks, which we want to never happen.

The mechanism we use to trigger getting called to check for state
transitions is with the data lowater event, previously used for signaling
data device extensions, now also for when FreeSpaceState thresholds are
crossed. This amounts to a fair amount of math to calculate these, and
then also to set the lowater event for the earlier of its two possible
trigger cases.

This commit also changes the amount that the data device grows from a
doubling of existing size to a fixed 1 GiB, and the meta device from
a doubling to adding another 16 MiB. This should let us track data dev
size more closely with its actual space needs.

This commit also adds and uses datablocks_to_sectors() and
sectors_to_datablocks(), which are also used in a few unrelated places.

Finally, we now call ThinPool::check() after creation or setup, since the
PoolState or PoolSpaceState may need immediate updating. e.g. if
PoolSpaceState is Warn and then the daemon restarts, we don't want to
wait for an event before updating PoolSpaceState to its proper value.

Thanks to @mulkieran for extensive review and code fixes.

Signed-off-by: Andy Grover <agrover@redhat.com>