Framework for issuing dbus signals based on engine events
Make a global engine event listener list.  The main loop uses the list to
register as a listener - the engine uses the list to send notifications.

Implement an EngineListener to translate engine events to the appropriate
dbus signals/handle as required.

Events implemented:
   BlockdevStateChanged
   FilesystemRenamed
   FilesystemUsedChanged
   PoolRenamed

Wrap Option<dbus::Path> in MaybeDbusPath.  This lets the conditional code
move inside the MaybeDbusPath struct definition, and keeps it out of the
rest of the Engine code.

Signed-off-by: Todd Gill <tgill@redhat.com>