Use thindev status for Filesystem::used()
Instead of using statvfs to get filesystem usage (which requires the
filesystem be mounted) use the thindev working status, which gives the
number of sectors used by the thindev. This is a little less granular
than statvfs(), but works on mounted and unmounted filesystems, which
avoids the temp-mount hijinks the previous code had.

Signed-off-by: Andy Grover <agrover@redhat.com>