Do not abort setting up filesystems if one fails
As discussed in previous commit, it's possible to be left with metadata
for a filesystem that was destroyed. Attempting to setup one of these
will fail. Instead of halting setup of all filesystems, emit a log message
but continue.

Also, attempt to remove the non-working metadata so this only happens
once.

Signed-off-by: Andy Grover <agrover@redhat.com>