Remove code that deletes md for failed thindevs, and FIXME
This will mean that we will try to setup the failing thindev on each
restart. This is probably ok for 1.0.

Once Filesystems have a State, we can keep these in the fs list with
State::Bad, and then maybe have the user choose to explicitly remove it?
Would that be a better way to handle this?

Signed-off-by: Andy Grover <agrover@redhat.com>