Replace "capacity" in method names with "size" and "allocated"
It may not always be clear what "capacity" means in any context, whether
it's the amount allocated (from me to someone else) or if it's the total
amount (the total amount I have, maybe allocated *from* someone else).

Try to clear this up by using "size" for methods that return the total
amount, and "allocated" for methods that refer to total amount given to
someone else.

Signed-off-by: Andy Grover <agrover@redhat.com>