devlinks: Do not return results
Although having the symbolic links in /dev/stratis being correct
, it's not as important than actually bringing up or taking down a
file system and/or a pool.  Furthermore, as the file system is accessible
to end users etc., we cannot control what happens to it.  Thus we
shouldn't fail other operations based on what does or doesn't work.
End users should be using FS UUID and not the /dev/mapper/<pool>/<fs name>
as these change when pool names and fs names change.

Signed-off-by: Tony Asleson <tasleson@redhat.com>