Change Blockdev::total_size() to ::size()
With the new nomenclature, size() is the method name for a function
returning the entire size of something, including unusable sectors.

Since this method is ultimately used to back a D-Bus property called
BlockdevPhysicalSize, I think this makes the most sense to return, rather
than usable size.

Interestingly, size() is already implemented in the impl BlockDev block.
Remove this but make BlockDev::size() return the entire size.

Signed-off-by: Andy Grover <agrover@redhat.com>