Store thinpool status in ThinPool
Set it on creation and set up.

Use the cached value in ThinPool::total_physical_used.
Change the return type of total_physical_used so that it returns a simple
value, not a Result, since none of the functions that it calls can return
an error. Change the return type of methods that invoke this method,
including Pool::total_physical_used. Change D-Bus layer property method
for pool's total physical size so that it can not return an error.

Cache the value of ThinPool::thin_pool_status field in ThinPool::check(),
which is the only place where the thin pool device's status is checked.

Signed-off-by: mulhern <amulhern@redhat.com>