Avoid overlapping ranges
I think the reason the compiler does not complain is
that it evaluates the matches in order, so there is
no ambiguity to it, even if the matches do overlap.

This fixes a bug, since the new state would be
set to good if the overall_used_pct was exactly
SPACE_WARN_PCT, and similar idea for warn and
SPACE_CRIT_PCT, which is not exactly as desired.

It depends on your viewpoint what you consider to
be the size of the error. I would say that in the case
of SPACE_CRIT_PCT it is a 20% error, because the
pool goes into Crit state at 4% instead of 5%, and
for SPACE_WARN_PCT it is only a 10% error, based
on similar reasoning.

Signed-off-by: mulhern <amulhern@redhat.com>