Rewrite test_thinpool_expand
Get rid of dependencies on constant values that used to represent constants
for the pool, but do not anymore. Use actual values for the pool instance
being tested instead.

Flush the buffer when necessary in order to ensure that the pool is actually
written to.

Verify that the status of the pool is as expected at various checkpoints.

Extend the thin pool by an amount which should always make the test run
correctly.

Signed-off-by: mulhern <amulhern@redhat.com>