README: Tell people how to get the correct rustfmt version
Since rustfmt is unstable from version to version (it changes formatting)
at least until rustfmt 1.0 is released, we are currently sticking to
rustfmt 0.6.1. This is kind of a pain to get, so document how to do so,
both on Fedora, and on other platforms, the latter through the use of
Rustup.

Signed-off-by: Andy Grover <agrover@redhat.com>