Call udev_settle() after failed fs create
If we are in the process of running mkfs on a new filesystem when the
pool is exhausted, we should clean up by destroying the new device and
returning an error. However, it appears udev jumps into motion based upon
mkfs.xfs writing its initial magic number, and this can result in EBUSY
when we try and remove the device.

Instead, in this error path, call udev_settle() to ensure udev doesn't
have the device open when we go to remove it.

fixes #1254

Signed-off-by: Andy Grover <agrover@redhat.com>